checkpoint for HilbertBtree implementation

+++ hyracks/hyracks-data/hyracks-data-std/src/main/java/edu/uci/ics/hyracks/data/std/primitive/DoublePointable.java
@@ -24,15 +24,15 @@ import edu.uci.ics.hyracks.data.std.api.IPointableFactory;
 
 public final class DoublePointable extends AbstractPointable implements IHashable, IComparable, INumeric {
 	private final static double machineEpsilon;
-	static {
-		float epsilon = 1.0f;
+	   static {
+	        double epsilon = 1.0d;
 
-        do {
-           epsilon /= 2.0f;
-        }
-        while ((float)(1.0 + (epsilon/2.0)) != 1.0);
-        machineEpsilon = epsilon;
-	}
+	        do {
+	           epsilon /= 2.0d;
+	        }
+	        while ((double)(1.0 + (epsilon/2.0)) != 1.0);
+	        machineEpsilon = epsilon;
+	    }
 	
     public static final ITypeTraits TYPE_TRAITS = new ITypeTraits() {
         private static final long serialVersionUID = 1L;
